requirements.txt
requests
python-qbittorrent
tabulate
urllib3
json